requirements.txt
fastapi
uvicorn
openai
pinecone
langchain-text-splitters
pymupdf
pillow
pytesseract
python-multipart
dotenv
langchain
langchain-openai
langchain-pinecone
langchain-core